* shopIndex, productIndex, productCategoryIndex, parentProductCategoryIndex, productTagIndex added * increments to bigIncrements in some tables * integer to bigInteger in some tables * products, product_categories, product_tags table added * E-commerce permissions added * shops_prefix Setting added * Money field added * Time field added * lang changed * templates for E-commerce added * sitemap routes for E-commerce added * ActionController.php * E-commerce functions in IndexController.php * Model::class changed * Resource:class changed * E-commerce policy added * E-commerce resources added * jsonProp function in helper.php changed * composer dependency "moneyphp/money" added

+++ database/migrations/2020_01_01_000000_create_portfolio_table.php
@@ -16,7 +16,7 @@ class CreatePortfolioTable extends Migration
     {
         Schema::create('portfolio', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('author_id');
+            $table->bigInteger('author_id');
             $table->string('title');
             $table->string('guard_name');
             $table->text('excerpt')->nullable();
