+++ database/migrations/2020_01_01_000000_create_posts_table.php
@@ -16,7 +16,7 @@ class CreatePostsTable extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('author_id');
+            $table->bigInteger('author_id');
             $table->string('title');
             $table->string('guard_name');
             $table->text('excerpt')->nullable();
